# Conflicts:
#	src/Core/UserBundle/Entity/User.php

+++ src/EBM/KMBundle/Controller/DocumentController.php
@@ -10,6 +10,8 @@ namespace EBM\KMBundle\Controller;
 
 
 use EBM\KMBundle\Entity\Document;
+use EBM\KMBundle\Entity\DocumentHistory;
+use EBM\KMBundle\Entity\DocumentRepository;
 use EBM\KMBundle\Entity\EvaluationDocument;
 use EBM\KMBundle\Entity\Post;
 use EBM\KMBundle\Entity\Topic;
@@ -17,6 +19,7 @@ use EBM\KMBundle\Form\DocumentType;
 use EBM\KMBundle\Form\EvaluationDocumentType;
 use EBM\KMBundle\Form\PostType;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
@@ -36,7 +39,7 @@ class DocumentController extends Controller
      */
     public function indexAction(){
 
-        $documents = $this->getDoctrine()->getRepository('EBMKMBundle:Document')->findAll();
+        $documents = $this->getDoctrine()->getRepository('EBMKMBundle:Document')->findBy(['active' => true]);
 
         return $this->render('EBMKMBundle:Documents:index.html.twig', array("documents" => $documents));
     }
@@ -52,6 +55,7 @@ class DocumentController extends Controller
     public function uploadAction(Request $request){
 
         $document = new Document();
+        $documentHistory = new DocumentHistory();
 
         // On récupère les différents tags pour les passer dans le formulaire.
         $tags = $this->getDoctrine()->getRepository('EBMKMBundle:Tag')->findAll();
@@ -61,30 +65,18 @@ class DocumentController extends Controller
         $form->handleRequest($request);
 
         if($form->isSubmitted() && $form->isValid()){
-            /** @var UploadedFile $file */
-            $file = $document->getFile();
 
-            $fileName = md5(uniqid()).'.'.$file->guessExtension();
-
-            $file->move(
-                $this->getParameter('files_directory'),
-                $fileName
-            );
-
-            // replace the file by its name
-            $document->setFile($fileName);
-
-            // The document's author is the current user
+            // Le créateur du document est l'utilisateur courrant.
             $user = $this->getUser();
-            $document->setAuthor($user);
+            $documentHistory->setAuthor($user);
+            $document->setHistory($documentHistory);
 
-            // Persist the newly created document
+            // On persiste le document et son historique nouvellement créé.
             $em = $this->getDoctrine()->getManager();
             $em->persist($document);
             $em->flush();
 
             return $this->redirect($this->generateUrl('ebmkm_document_index'));
-
         }
 
         return $this->render('EBMKMBundle:Documents:upload.html.twig', array(
@@ -109,19 +101,27 @@ class DocumentController extends Controller
         $user = $this->getUser();
 
         /*
-         *  Cette partie charge le fil de commentaires sur le document.
-         * S'il n'y en a pas, il sera créé avec le premier commentaire.
+         * Cette partie charge le fil de commentaires sur le document.
+         * S'il n'y en a pas, il sera créé lorsque l'utilisateur commentera pour la première fois.
          * Une fois le message posté, la page est raffraichie.
+         *
+         * Le topic créé aura le nom du document suivi de '- Commentaires'.
+         * Une description est générée automatiquement.
+         *
+         * //TODO : Bouger le texte de la description dans un fichier avec tous les textes.
+         *
          */
-        if($document->getCommentTopic()){
-            $topic = $document->getCommentTopic();
+        if($document->getHistory()->getCommentTopic()){
+            $topic = $document->getHistory()->getCommentTopic();
         }
         else{
             $topic = new Topic();
             $topic
-                ->setTitle($document->getName())
+                ->setTitle($document->getName() . ' - Commentaires')
                 ->setCreator($user);
-            $document->setCommentTopic($topic);
+            $topic->setDescription("Ceci est le fil de discussion relatif au document ' " . $document->getName() . '.');
+
+            $document->getHistory()->setCommentTopic($topic);
         }
 
         $post = new Post();
@@ -131,6 +131,9 @@ class DocumentController extends Controller
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
 
+        /*
+         * Le fil de commentaire n'est créé et complété par les posts que si le nouveau post est bien formé.
+         */
         if($form->isSubmitted() && $form->isValid())
         {
             $em->persist($topic);
@@ -147,12 +150,12 @@ class DocumentController extends Controller
          * L'évaluation d'un document se fait via un formulaire utilisant un slider.
          * Une fois ce formulaire posté, la page est raffraichie.
          */
-        if(sizeof($document->getEvaluations()) > 0){
+        if(sizeof($document->getHistory()->getEvaluations()) > 0){
             $moyenne = 0;
-            foreach ($document->getEvaluations() as $evaluation){
+            foreach ($document->getHistory()->getEvaluations() as $evaluation){
                 $moyenne += $evaluation->getValue();
             }
-            $moyenne = $moyenne/sizeof($document->getEvaluations());
+            $moyenne = $moyenne/sizeof($document->getHistory()->getEvaluations());
         }
         else{
             $moyenne = -1;
@@ -161,12 +164,15 @@ class DocumentController extends Controller
         $personalEvaluation = $this->getDoctrine()->getRepository('EBMKMBundle:EvaluationDocument')
             ->findBy(['author' => $this->getUser(), 'document' => $document]);
 
+        /*
+         * Evaluation du document par l'utilisateur courrant.
+         */
         $evaluation = new EvaluationDocument();
         $evaluationForm = $this->createForm(EvaluationDocumentType::class, $evaluation);
         $evaluationForm->handleRequest($request);
         if($evaluationForm->isSubmitted() && $evaluationForm->isValid()){
             $evaluation->setAuthor($user);
-            $evaluation->setDocument($document);
+            $evaluation->setDocument($document->getHistory());
             $em->persist($evaluation);
             $em->flush();
             return $this->redirectToRoute('ebmkm_document_detail', array('id' => $document->getId()));
@@ -192,33 +198,67 @@ class DocumentController extends Controller
      * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
      */
     public function updateAction($id, Request $request){
-        $document = $this->getDoctrine()->getRepository('EBMKMBundle:Document')->find($id);
+        /**
+         * On récupère un clone de la dernière version du document.
+         *
+         * @var Document $document
+         */
+        $oldDocument = $this->getDoctrine()->getRepository('EBMKMBundle:Document')->find($id);
+        $updateDocument = clone $oldDocument;
+
         $tags = $this->getDoctrine()->getRepository('EBMKMBundle:Tag')->findAll();
 
-        $deleteForm = $this->createDeleteForm($document);
-        $editForm = $this->createForm(DocumentType::class, $document, array(
+        /*
+         * On récupère tout d'abord le chemin du document.
+         * On peut ensuite récupérer le document en lui-même, et l'affecter au champ 'File' du Document.
+         */
+        $helper = $this->get('vich_uploader.templating.helper.uploader_helper');
+        $path = $helper->asset($oldDocument, 'file');
+        $kernel_root_dir = $this->getParameter('kernel.root_dir');
+        $file = new File\File( $kernel_root_dir . '/../web' . $path);
+        $updateDocument->setFile($file);
+
+        $deleteForm = $this->createDeleteForm($updateDocument);
+        $editForm = $this->createForm(DocumentType::class, $updateDocument, array(
             'tags' => $tags
         ));
         $editForm->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+        /*
+         * Si les données sont validées et que l'utilisateur connecté est bien l'auteur du document :
+         * - On récupère le nouveau formulaire, et on applique les modifications.
+         * - On rajoute le nouveau document dans l'historique.
+         */
+        if ($editForm->isSubmitted()
+            && $editForm->isValid()
+            && $this->getUser() == $oldDocument->getHistory()->getAuthor()) {
 
-            return $this->redirectToRoute('ebmkm_document_detail', array('id' => $document->getId()));
+            $oldDocument->setActive(false);
+            $updateDocument->setDate(new \DateTime());
+
+            if(!$updateDocument->getFile()){
+                $updateDocument->setFile($oldDocument->getFile());
+            }
+
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($updateDocument);
+            $em->flush();
+
+            return $this->redirectToRoute('ebmkm_document_detail', array('id' => $updateDocument->getId()));
         }
 
         return $this->render('EBMKMBundle:Documents:update.html.twig', array(
-            'document' => $document,
+            'document' => $updateDocument,
             'edit_form' => $editForm->createView(),
             'delete_form' => $deleteForm->createView(),
         ));
-
     }
 
     /**
-     * Deletes a document.
-     *
-     * @Method("DELETE")
+     * Deletes a document, all its history, grades and comments.
+     * @param Request $request
+     * @param $id
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
      */
     public function deleteAction(Request $request, $id)
     {
@@ -228,12 +268,11 @@ class DocumentController extends Controller
         $form = $this->createDeleteForm($document);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
+        if ($form->isSubmitted() && $form->isValid() && $this->getUser() == $document->getHistory()->getAuthor()) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($document);
             $em->flush($document);
         }
-
 
         return $this->redirectToRoute('ebmkm_document_index');
     }
@@ -248,10 +287,10 @@ class DocumentController extends Controller
     private function createDeleteForm(Document $document)
     {
         return $this->createFormBuilder()
+            ->add('submit', SubmitType::class, array('label' => 'Supprimer le document'))
             ->setAction($this->generateUrl('ebmkm_document_delete', array('id' => $document->getId())))
             ->setMethod('DELETE')
-            ->getForm()
-            ;
+            ->getForm();
     }
 
 }
